project/requirements.txt
asgiref==3.8.1
Django==3.2.25
djangorestframework==3.15.1
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.11.0
